refactored the music widget in homepage to new file
