+++ lib/Custom Widgets/swipecard.dart
@@ -1,11 +1,12 @@
-import 'package:dateme/Classes/Person.dart';
-import 'package:dateme/constants.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_swipable/flutter_swipable.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 
 //modules
 import 'package:dateme/Classes/personList.dart';
+import 'musicWidgets.dart';
+import 'package:dateme/Classes/Person.dart';
+import 'package:dateme/constants.dart';
 
 class SwipeCard extends StatelessWidget {
   SwipeCard(
@@ -13,6 +14,7 @@ class SwipeCard extends StatelessWidget {
       required this.profilePicture,
       required this.colour,
       required this.firstName,
+      required this.lastName,
       required this.age,
       required this.bio})
       : super(key: key);
@@ -20,6 +22,7 @@ class SwipeCard extends StatelessWidget {
   NetworkImage profilePicture;
   Color colour;
   String firstName;
+  String lastName;
   int age;
   String bio;
 
@@ -35,7 +38,8 @@ class SwipeCard extends StatelessWidget {
                 borderRadius: BorderRadius.vertical(top: Radius.circular(20))),
             isScrollControlled: true,
             context: context,
-            builder: (context) => buildSheet(context));
+            builder: (context) =>
+                BuildSheet(firstname: firstName, bio: bio, age: age));
         // Navigator.push(context, Drawer())
       },
       child: Swipable(
@@ -69,7 +73,19 @@ class SwipeCard extends StatelessWidget {
       ),
     );
   }
+}
 
+class BuildSheet extends StatefulWidget {
+  String firstname, bio;
+  int age;
+
+  BuildSheet({required this.firstname, required this.bio, required this.age});
+
+  @override
+  _BuildSheetState createState() => _BuildSheetState();
+}
+
+class _BuildSheetState extends State<BuildSheet> {
   Widget makeDismissable(
           {required Widget child, required BuildContext context}) =>
       GestureDetector(
@@ -81,161 +97,70 @@ class SwipeCard extends StatelessWidget {
         ),
       );
 
-  buildSheet(context) => makeDismissable(
-        context: context,
-        child: DraggableScrollableSheet(
-          maxChildSize: 0.7,
-          initialChildSize: 0.3,
-          builder: (_, controller) => Container(
-            decoration: BoxDecoration(
-                color: Colors.white,
-                borderRadius: BorderRadius.vertical(top: Radius.circular(20))),
-            child: ListView(
-              controller: controller,
-              children: [
-                Padding(
-                  padding: const EdgeInsets.fromLTRB(20, 15, 0, 10),
-                  child: Text(
-                    'First Name, 21',
-                    style: kboldNameStyle,
-                  ),
-                ),
-                Divider(),
-                const Padding(
-                  padding: EdgeInsets.all(8.0),
-                  child: Text(
-                    "About Me",
-                    style: TextStyle(
-                        color: Colors.orange,
-                        fontSize: 20,
-                        fontWeight: FontWeight.bold,
-                        fontStyle: FontStyle.italic),
-                  ),
-                ),
-                const Padding(
-                  padding:
-                      EdgeInsets.symmetric(horizontal: 10.0, vertical: 15.0),
-                  child: Text(
-                      'Im a photgrapher and a model. Just want someone to hang out with.'),
-                ),
-                Divider(), //customize the divider
-                customInterestCard(
-                    interestText: "Interest 1", colour: Colors.red),
-                customInterestCard(
-                    interestText: "Interest 2", colour: Colors.orange),
-                customInterestCard(
-                    interestText: "Interest 3", colour: Colors.yellow),
-                Divider(),
-                customMusicCard(),
-                Center(
-                  child: IconButton(
-                      onPressed: () {
-                        Navigator.of(context).pop();
-                      },
-                      icon: Icon(Icons.close, size: 30, color: Colors.orange)),
-                )
-              ],
-            ),
-          ),
-        ),
-      );
-}
-
-class customMusicCard extends StatelessWidget {
-  const customMusicCard({Key? key}) : super(key: key);
-
   @override
   Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.symmetric(horizontal: 5.0),
-      child: Card(
-          //shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(20)),
-          child: Container(
-        child: Column(
-          children: [
-            Padding(
-              padding: const EdgeInsets.all(5.0),
-              child: Row(
-                children: [
-                  Flexible(child: musicTrackCard()),
-                ],
+    return makeDismissable(
+      context: context,
+      child: DraggableScrollableSheet(
+        maxChildSize: 0.7,
+        initialChildSize: 0.3,
+        builder: (_, controller) => Container(
+          decoration: BoxDecoration(
+              color: Colors.white,
+              borderRadius: BorderRadius.vertical(top: Radius.circular(20))),
+          child: ListView(
+            controller: controller,
+            children: [
+              Padding(
+                padding: const EdgeInsets.fromLTRB(20, 15, 0, 10),
+                child: Text(
+                  '${widget.firstname}, ${widget.age}',
+                  style: kboldNameStyleBlack,
+                ),
               ),
-            ),
-            Padding(
-              padding:
-                  const EdgeInsets.symmetric(vertical: 2.5, horizontal: 10),
-              child: Row(
-                children: [
-                  Flexible(flex: 3, child: customMessageButton()),
-                  Flexible(
-                      flex: 1,
-                      child: TextButton(
-                        style: TextButton.styleFrom(
-                            backgroundColor: Colors.white,
-                            shape: CircleBorder()),
-                        onPressed: () {},
-                        child: Padding(
-                          padding: const EdgeInsets.all(5.0),
-                          child: Icon(FontAwesomeIcons.userPlus,
-                              color: Colors.black),
-                        ),
-                      ))
-                ],
+              Divider(),
+              const Padding(
+                padding: EdgeInsets.all(8.0),
+                child: Text(
+                  "About Me",
+                  style: TextStyle(
+                      color: Colors.orange,
+                      fontSize: 20,
+                      fontWeight: FontWeight.bold,
+                      fontStyle: FontStyle.italic),
+                ),
               ),
-            )
-          ],
+              Padding(
+                padding: EdgeInsets.symmetric(horizontal: 10.0, vertical: 15.0),
+                child: Text('${widget.bio}'),
+              ),
+              Divider(), //customize the divider
+              customInterestCard(
+                  interestText: "Interest 1", colour: Colors.red),
+              customInterestCard(
+                  interestText: "Interest 2", colour: Colors.orange),
+              customInterestCard(
+                  interestText: "Interest 3", colour: Colors.yellow),
+              Divider(),
+              customMusicCard(),
+              Center(
+                child: IconButton(
+                    onPressed: () {
+                      Navigator.of(context).pop();
+                    },
+                    icon: Icon(Icons.close, size: 30, color: Colors.orange)),
+              )
+            ],
+          ),
         ),
-        height: MediaQuery.of(context).size.height * 0.4,
-        decoration: BoxDecoration(
-            color: Colors.green, borderRadius: BorderRadius.circular(20.0)),
-      )),
+      ),
     );
   }
 }
 
 //custom card for the music tracks
-class musicTrackCard extends StatelessWidget {
-  const musicTrackCard({
-    Key? key,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Card(
-      shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(20.0)),
-      child: Container(
-        decoration: BoxDecoration(
-            color: Colors.red, borderRadius: BorderRadius.circular(20.0)),
-        width: MediaQuery.of(context).size.width * 0.9,
-        height: MediaQuery.of(context).size.height * 0.3,
-      ),
-    );
-  }
-}
 
 //message button on the bottom sheet
-class customMessageButton extends StatelessWidget {
-  const customMessageButton({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return GestureDetector(
-      onTap: () => print("messaging person"),
-      child: Container(
-        child: Center(
-            child: Text('Message',
-                style: TextStyle(
-                    color: kpinkColor,
-                    fontSize: 15.0,
-                    fontWeight: FontWeight.w400))),
-        height: MediaQuery.of(context).size.height * 0.05,
-        width: MediaQuery.of(context).size.width * 0.7,
-        decoration: BoxDecoration(
-            color: Colors.white, borderRadius: BorderRadius.circular(20.0)),
-      ),
-    );
-  }
-}
 
 class customInterestCard extends StatelessWidget {
   String interestText;
@@ -275,6 +200,7 @@ class _SwipeStackState extends State<SwipeStack> {
     SwipeCard(
       colour: Colors.red,
       firstName: 'Nanamie',
+      lastName: 'Sakamoto',
       age: 21,
       bio: 'I\'m an independent woman who don\'t need no man (jk date me pls).',
       profilePicture: NetworkImage(
